requirements.txt
# requirements.txt
# Core
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# AI/ML
langgraph==0.0.20
langchain-google-genai==1.0.0
google-generativeai==0.3.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Text Processing
PyPDF2==3.0.1
python-docx==0.8.11
docx2txt==0.8
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1
scikit-learn==1.3.0

# Data Analysis
pandas==2.1.0
numpy==1.24.3
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2

# Database
sqlalchemy==2.0.21
alembic==1.12.1
redis==5.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-bcrypt==0.4.0
cryptography==41.0.5
pyjwt==2.8.0

# API
httpx==0.25.1
requests==2.31.0

# Email
aiofiles==23.2.1
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
typing-extensions==4.8.0
python-dateutil==2.8.2
tqdm==4.66.1
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.18.0
psutil==5.9.6